rename convnet to dynamic_asinh_convnet
