+++ case_studies/dc2_cataloging/utils/dynamic_asinh_convnet.py
